+++ R/ensemble_sdm.R
@@ -33,12 +33,15 @@
 #' \code{weighted_average} is the same average, but weighted by a metric, which needs to be
 #' set using argument \code{metric} (see \link{mean_validation_metrics} for the metrics available).
 #' \code{committee_average} is the committee average, as known as majority rule, where predictions
-#' are binarized and then a mean is obtained. To binarize predictions, user can set a custom function
-#' in the \code{fun} argument to calculate a threshold for each model. Standardly, the
+#' are binarized and then a mean is obtained. To binarize predictions, user can set a custom
+#' function in the \code{fun} argument to calculate a threshold for each model. Standardly, the
 #' committee average uses the \code{caret::thresholder} function to find the threshold that
-#' maximizes the sum of sensitivity and specificity (through \code{caretSDM:::.MaxSeSp}). The custom
-#' function must use the argument \code{mod}, which is the model output from caret package (see
-#' \code{\link{get_models}}) and must return a \code{numeric} value (see example).
+#' maximizes the sum of sensitivity and specificity (through \code{caretSDM:::.MaxSeSp}).
+#' Custom function (\code{fun}) must use the argument \code{mod}, which is the model output from
+#' caret package (see \code{\link{get_models}}) and must return a \code{numeric} value (see example).
+#' \code{method} can also be set to a custom function, which must receive the argument pred_mat,
+#' which is a matrix of predictions (columns are models and rows are cells) and return a vector of
+#' predictions (one value per cell). See the \code{median} example below for a custom function.
 #'
 #' \code{get_predictions} returns the \code{list} of all predictions to all scenarios, all species,
 #' all algorithms and all repetitions. Useful for those who wish to implement their own ensemble
@@ -112,6 +115,11 @@
 #'    if (length(th) > 1) mean(th) else th
 #' }
 #'
+#' # Example from a custom function to obtain ensembles using the median instead of the mean:
+#' median_ensemble <- function(pred_mat) {
+#'  apply(pred_mat, 1, median, na.rm = TRUE)
+#' }
+#'
 #' @importFrom caret thresholder
 #' @import checkCLI
 #'
@@ -159,6 +167,7 @@ ensemble_sdm <- function(m, scen = NULL, method = "average", metric = NULL, fun 
     } else {
       #perform checks on the custom function
       fun_name <- deparse(substitute(fun))
+      assert_function_cli(fun, args = "mod")
       message (paste0("Binarization Function: ", fun_name))
     }
     bin_thresholds <- lapply(names(y$models),
@@ -188,7 +197,11 @@ ensemble_sdm <- function(m, scen = NULL, method = "average", metric = NULL, fun 
       l <- list()
 
       if(is.function(method)){
-        l[[]] <- method(pred_mat)
+        assert_function_cli(method, args = "pred_mat")
+        ens <- method(pred_mat)
+        assert_numeric_cli(ens, len = nrow(pred_mat))
+        assert_vector_cli(ens, len = nrow(pred_mat))
+        l[[method_name]] <- ens
       } else {
         if("average" %in% method){
           ens <- rowMeans(pred_mat, na.rm = TRUE)